client-py/requirements.txt
requests>=2.31.0
websockets>=11.0.2
python-dotenv>=1.0.0
